Copy+paste `buildProperties.php` logic into controller
